+++ Modules/Intent.Modules.ValueObjects/Templates/ValueObjectBase/ValueObjectBaseTemplate.cs
@@ -4,7 +4,7 @@ using Intent.Modules.Common.Templates;
 using Intent.RoslynWeaver.Attributes;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
-[assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpStringInterpolationTemplate", Version = "1.0")]
+[assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpStringInterpolationTemplate", Version= "1.0")]
 
 namespace Intent.Modules.ValueObjects.Templates.ValueObjectBase
 {