+++ Modules/Intent.Modules.Blazor.HttpClients.Dtos.FluentValidation/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 1.2.2
 
+- Improvement: Updated module icon
 - Improvement: `ConfigureValidationRules` method is no longer be created on _validators_ if it would be created empty.
 
 ### Version 1.2.1